test: cleaned up the reflection for trying all components once
